backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
pdf2image==1.17.0
paddleocr==2.7.3
paddlepaddle==2.6.2
pytesseract==0.3.10
Pillow==10.3.0
openpyxl==3.1.2
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
